requirements.txt
anthropic>=0.18.0
exa-py>=1.0.0
streamlit>=1.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
